smashed class Man into classes and made almost a worker
