+++ src/parse/Race.java
@@ -1,8 +1,8 @@
-public class Race extends Man<Race> {
+public class Race  {
     public String RaceName;
+    public String race;
+    public String sex;
+    public int paraped;
+    public boolean pidr;
 
-
-    public static Race createContents() {
-        return new Race();
-    }
 }
